requirements.txt
streamlit>=1.47.0
mysql-connector-python>=8.0.33
pandas>=1.5.0
plotly>=5.15.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
Pillow>=9.5.0
boto3>=1.26.0
awscli>=1.27.0
